Fixes
I had done some fixes to some of the pages such as the malito email sender and the paragraph boxes. Within the courses tab I had made the two boxes to allow you to go to one place to other pages such as the part time courses page and the full time courses page.