Finished Coding
I finished arranging the course link boxes in the course tabs. After that I started adding in the codes for header, footer and title in both the full and part time courses. After finishing the coding in the courses page I went back to the contact page and added a picture of the map of the school. I changed the titles from documents to their respective name pages and lastly fixing the paragraph in the index page. I had also added task 1 and 2's pdfs within the code